Fully move key and mouse press/release methods from Game into GameInputReceiver.
