+++ rts/Game/Game.cpp
@@ -26,11 +26,8 @@
 #include "UnsyncedGameCommands.h"
 #include "Game/Players/Player.h"
 #include "Game/Players/PlayerHandler.h"
-#include "Game/UI/KeyCodes.h"
-#include "Game/UI/KeySet.h"
 #include "Game/UI/PlayerRoster.h"
 #include "Game/UI/PlayerRosterDrawer.h"
-#include "Game/UI/ScanCodes.h"
 #include "Game/UI/UnitTracker.h"
 #include "ExternalAI/AILibraryManager.h"
 #include "ExternalAI/EngineOutHandler.h"
@@ -215,8 +212,6 @@ CR_REG_METADATA(CGame, (
 	CR_MEMBER(luaGCControl),
 
 	CR_IGNORED(jobDispatcher),
-	CR_IGNORED(curKeyCodeChain),
-	CR_IGNORED(curScanCodeChain),
 	CR_IGNORED(worldDrawer),
 	CR_IGNORED(saveFileHandler),
 	CR_IGNORED(gameInputReceiver),
@@ -1108,150 +1103,9 @@ void CGame::ResizeEvent()
 }
 
 
-int CGame::KeyPressed(int keyCode, int scanCode, bool isRepeat)
-{
-	RECOIL_DETAILED_TRACY_ZONE;
-	if (!gameOver && !isRepeat)
-		playerHandler.Player(gu->myPlayerNum)->currentStats.keyPresses++;
-
-	const CKeySet kc(keyCode, CKeySet::KSKeyCode);
-	const CKeySet ks(scanCode, CKeySet::KSScanCode);
-
-	curKeyCodeChain.push_back(kc, spring_gettime(), isRepeat);
-	curScanCodeChain.push_back(ks, spring_gettime(), isRepeat);
-
-	lastActionList = keyBindings.GetActionList(curKeyCodeChain, curScanCodeChain);
-
-	if (RmlGui::ProcessKeyPressed(keyCode, scanCode, isRepeat))
-		return 0;
-
-	if (gameTextInput.ConsumePressedKey(keyCode, scanCode, lastActionList))
-		return 0;
-
-	if (luaInputReceiver->KeyPressed(keyCode, scanCode, isRepeat))
-		return 0;
-
-
-	// try the input receivers
-	for (CInputReceiver* recv: CInputReceiver::GetReceivers()) {
-		if (recv != nullptr && recv->KeyPressed(keyCode, scanCode, isRepeat))
-			return 0;
-	}
-
-	// try our list of actions
-	for (const Action& action: lastActionList) {
-		if (ActionPressed(action, isRepeat)) {
-			return 0;
-		}
-	}
-
-	// maybe a widget is interested?
-	if (luaUI != nullptr) {
-		for (const Action& action: lastActionList) {
-			luaUI->GotChatMsg(action.rawline, false);
-		}
-	}
-
-	if (luaMenu != nullptr) {
-		for (const Action& action: lastActionList) {
-			luaMenu->GotChatMsg(action.rawline, false);
-		}
-	}
-
-	return 0;
-}
-
-
-int CGame::KeyReleased(int keyCode, int scanCode)
-{
-	RECOIL_DETAILED_TRACY_ZONE;
-	if (RmlGui::ProcessKeyReleased(keyCode, scanCode))
-		return 0;
-
-	if (gameTextInput.ConsumeReleasedKey(keyCode, scanCode))
-		return 0;
-
-	// update actionlist for lua consumer
-	lastActionList = keyBindings.GetActionList(keyCode, scanCode);
-
-	if (luaInputReceiver->KeyReleased(keyCode, scanCode))
-		return 0;
-
-	// try the input receivers
-	for (CInputReceiver* recv: CInputReceiver::GetReceivers()) {
-		if (recv != nullptr && recv->KeyReleased(keyCode, scanCode)) {
-			return 0;
-		}
-	}
-
-	for (const Action& action: lastActionList) {
-		if (ActionReleased(action))
-			return 0;
-	}
-
-	return 0;
-}
-
 CInputReceiver* CGame::GetInputReceiver()
 {
 	return &gameInputReceiver;
-}
-
-bool CGame::MousePress(int x, int y, int button)
-{
-	int keyCode = CKeyCodes::GetMouseButtonSymbol(button);
-	int scanCode = CScanCodes::GetMouseButtonSymbol(button);
-
-	const CKeySet kc(keyCode, CKeySet::KSKeyCode);
-	const CKeySet ks(scanCode, CKeySet::KSScanCode);
-	bool isRepeat = false;
-
-	const auto now = spring_gettime();
-	curKeyCodeChain.push_back(kc, now, isRepeat);
-	curScanCodeChain.push_back(ks, now, isRepeat);
-
-	lastActionList = keyBindings.GetActionList(curKeyCodeChain, curScanCodeChain);
-
-	// try our list of actions
-	for (const Action& action: lastActionList) {
-		if (ActionPressed(action, isRepeat)) {
-			return true;
-		}
-	}
-
-	// maybe a widget is interested?
-	// allowing all listeners to process for backwards compatibility.
-	bool handled = false;
-
-	if (luaUI != nullptr) {
-		for (const Action& action: lastActionList) {
-			handled |= luaUI->GotChatMsg(action.rawline, false);
-		}
-	}
-
-	if (luaMenu != nullptr) {
-		for (const Action& action: lastActionList) {
-			handled |= luaMenu->GotChatMsg(action.rawline, false);
-		}
-	}
-
-	return handled;
-}
-
-bool CGame::MouseRelease(int x, int y, int button)
-{
-	int keyCode = CKeyCodes::GetMouseButtonSymbol(button);
-	int scanCode = CScanCodes::GetMouseButtonSymbol(button);
-
-	// update actionlist for lua consumer
-	lastActionList = keyBindings.GetActionList(keyCode, scanCode);
-
-	for (const Action& action: lastActionList) {
-		if (ActionReleased(action))
-			return true;
-	}
-
-	return false;
 }
 
 int CGame::KeyMapChanged()
