+++ rts/Game/Game.h
@@ -10,7 +10,6 @@
 #include "GameController.h"
 #include "GameJobDispatcher.h"
 #include "GameInputReceiver.h"
-#include "Game/UI/KeySet.h"
 #include "Game/Action.h"
 #include "Rendering/WorldDrawer.h"
 #include "System/UnorderedMap.hpp"
@@ -107,6 +106,8 @@ public:
 	void SetDrawMode(GameDrawMode mode) { gameDrawMode = mode; }
 	GameDrawMode GetDrawMode() const { return gameDrawMode; }
 
+	bool ActionPressed(const Action& action, bool isRepeat);
+	bool ActionReleased(const Action& action);
 private:
 	bool Draw() override;
 	bool Update() override;
@@ -120,13 +121,6 @@ private:
 	/// Format and display a chat message received over network
 	void HandleChatMsg(const ChatMessage& msg);
 
-	/// Called when a key is released by the user
-	int KeyReleased(int keyCode, int scanCode) override;
-	/// Called when the key is pressed by the user (can be called several times due to key repeat)
-	int KeyPressed(int keyCode, int scanCode, bool isRepeat) override;
-
-	bool MousePress(int x, int y, int button) override;
-	bool MouseRelease(int x, int y, int button) override;
 	CInputReceiver* GetInputReceiver() override;
 
 	/// Called when the keymap changes (language or keyboard switch)
@@ -135,8 +129,6 @@ private:
 	int TextInput(const std::string& utf8Text) override;
 	int TextEditing(const std::string& utf8Text, unsigned int start, unsigned int length) override;
 
-	bool ActionPressed(const Action& action, bool isRepeat);
-	bool ActionReleased(const Action& action);
 	/// synced actions (received from server) go in here
 	void ActionReceived(const Action& action, int playerID);
 
@@ -218,9 +210,6 @@ public:
 
 private:
 	JobDispatcher jobDispatcher;
-
-	CTimedKeyChain curKeyCodeChain;
-	CTimedKeyChain curScanCodeChain;
 
 	CWorldDrawer worldDrawer;
